requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.8
flask_socketio==4.3.0
neo4j-driver==1.7.6
requests==2.22.0
eventlet
gunicorn
